refactor: dropdown, slider and toggle components
