+++ src/Slider/Slider.jsx
