--- src/dropdown/index.jsx
@@ -1,3 +1,0 @@
-import Dropdown from './dropdown.jsx'
-
-export default Dropdown
