--- src/input/index.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import clsx from 'clsx'
-
-const Input = ({ className, ...props }) => {
-  return <input className={clsx(className, 'lexia')} {...props} />
-}
-
-export default Input
