--- src/slider/index.jsx
@@ -1,3 +1,0 @@
-import Slider from './slider.jsx'
-
-export default Slider
