--- src/toggle/index.jsx
@@ -1,3 +1,0 @@
-import Toggle from './toggle.jsx'
-
-export default Toggle
